requirements.txt
pandas
pytest
django
pre-commit
djangorestframework
flake8
django-filter
djangorestframework-api-key
